Extract detector load/run into detector.ts; unload SAM 3 before loading a new detector
Addresses two review comments on worker.ts:

- Move the two detector paths (SAM 3 text-promptable + closed-vocabulary)
  out of worker.ts into a dedicated detector.ts, mirroring the existing
  sam3 module split. Exposes loadDetector / runDetector / isDetectorReady
  (plus the shared patchProcessorSize). The classifier stays in worker.ts
  for a later PR.
- When switching to a closed-vocabulary detector, unload SAM 3 sessions
  before allocating the new detector (and free any closed-vocab detector
  before loading SAM 3), so we never hold both in memory at once — reduces
  peak memory / OOM risk.
