+++ nachet-mini/src/inference/worker.ts
@@ -2,7 +2,6 @@
 
 import {
   AutoProcessor,
-  AutoModelForObjectDetection,
   AutoModelForImageClassification,
   RawImage,
   Tensor,
@@ -12,7 +11,12 @@ import {
 } from "@huggingface/transformers";
 import type { ModelConfig, WorkerInMessage, WorkerOutMessage } from "./models";
 import type { InferenceResult, InferenceBox } from "@common/types";
-import { loadSam3, runSam3, unloadSam3 } from "./sam3";
+import {
+  loadDetector,
+  runDetector,
+  isDetectorReady,
+  patchProcessorSize,
+} from "./detector";
 
 // ---------------------------------------------------------------------------
 // Environment
@@ -25,16 +29,6 @@ env.allowRemoteModels = true;
 // static server returns an HTML 404 page which transformers.js tries to parse
 // as JSON and fails. Disable local model lookup in production.
 env.allowLocalModels = import.meta.env.DEV;
-
-// ---------------------------------------------------------------------------
-// Types for post-processing output
-// ---------------------------------------------------------------------------
-
-interface PostProcessedDetection {
-  boxes: number[][];
-  classes: number[];
-  scores: number[];
-}
 
 // ---------------------------------------------------------------------------
 // Worker-specific helpers
@@ -94,44 +88,9 @@ const cropRegion = async (
 };
 
 // ---------------------------------------------------------------------------
-// Processor size patching
-// ---------------------------------------------------------------------------
-
-/**
- * Some HuggingFace models (e.g. RT-DETR from cfia-ai-lab) use
- * `{ max_height, max_width }` in their preprocessor_config.json `size` field.
- * transformers.js doesn't support this format, so we convert it to
- * `{ longest_edge }` which preserves aspect ratio.
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const patchProcessorSize = (processor: any): void => {
-  // AutoProcessor wraps an image_processor; try both paths
-  const imageProcessor = processor?.image_processor ?? processor;
-  if (!imageProcessor?.size) {
-    console.log("[worker] No image processor size to patch");
-    return;
-  }
-
-  const size = imageProcessor.size;
-  console.log("[worker] Processor size config:", JSON.stringify(size));
-
-  if (size.max_height !== undefined && size.max_width !== undefined) {
-    const longest = Math.min(size.max_height, size.max_width);
-    console.log(
-      `[worker] Patching processor size: {max_height: ${size.max_height}, max_width: ${size.max_width}} → {longest_edge: ${longest}}`,
-    );
-    imageProcessor.size = { longest_edge: longest };
-  }
-};
-
-// ---------------------------------------------------------------------------
 // Pipeline state
 // ---------------------------------------------------------------------------
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-let detectorModel: any = null;
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-let detectorProcessor: any = null;
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 let classifierModel: any = null;
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -176,8 +135,6 @@ addEventListener("message", async (event: MessageEvent) => {
   if (data.type === "load-models") {
     const config = data.config;
     const device = await getDevice();
-    // WebGPU has precision issues with detection models — use WASM for detector
-    const detectorDevice: DeviceType = "wasm";
     const classifierDevice = device;
     const progressDetector = makeProgressCallback("detector");
     const progressClassifier = makeProgressCallback("classifier");
@@ -185,12 +142,7 @@ addEventListener("message", async (event: MessageEvent) => {
     try {
       send({ type: "status", status: "loading-model" });
 
-      console.log(
-        "[worker] Loading detector model:",
-        config.detectorModel,
-        "device:",
-        detectorDevice,
-      );
+      console.log("[worker] Loading detector model:", config.detectorModel);
       console.log(
         "[worker] Loading classifier model:",
         config.classifierModel,
@@ -198,58 +150,21 @@ addEventListener("message", async (event: MessageEvent) => {
         classifierDevice,
       );
 
-      // Detector loading: two paths.
-      //
-      // 1. text-promptable-segmentation (SAM 3) — orchestrate 3 ONNX files
-      //    via raw onnxruntime-web. The transformers.js AutoModel APIs
-      //    can't represent this kind of multi-component, text-conditioned
-      //    detector. Delegated to the sam3 module.
-      //
-      // 2. object-detection (default) — single-file model loaded through
-      //    transformers.js's AutoModelForObjectDetection. The original
-      //    closed-vocabulary path (RT-DETR, DETR).
-      if (config.detectorKind === "text-promptable-segmentation") {
-        console.log("[worker] Loading SAM 3 detector via sam3 module");
-        // SAM 3's three components — vision encoder, text encoder, decoder —
-        // are loaded inside the sam3 module. We forward progress events.
-        await loadSam3(config, (info) => {
+      // Detector loading is delegated to the detector module, which handles
+      // both the SAM 3 (text-promptable) and closed-vocabulary paths and frees
+      // the other kind's memory before loading its own.
+      await loadDetector(config, {
+        transformersProgress: progressDetector as unknown as (
+          info: unknown,
+        ) => void,
+        sam3Progress: (info) => {
           send({
             type: "model-progress",
             name: `detector: ${info.name}`,
             progress: info.progress,
           });
-        });
-        // Leave detectorModel/detectorProcessor as null — the SAM 3 code
-        // path doesn't go through them. The classifier still loads below.
-        detectorModel = null;
-        detectorProcessor = null;
-      } else {
-        // Closed-vocabulary detector path (existing behavior).
-        const [detProc, detMod] = await Promise.all([
-          AutoProcessor.from_pretrained(config.detectorModel),
-          AutoModelForObjectDetection.from_pretrained(config.detectorModel, {
-            device: detectorDevice,
-            dtype: "fp32" as const,
-            model_file_name: config.detectorModelFileName ?? "model",
-            progress_callback: progressDetector as unknown as (
-              progress: unknown,
-            ) => void,
-          }),
-        ]);
-
-        console.log("[worker] Detector loaded, patching processor...");
-        patchProcessorSize(detProc);
-        detectorProcessor = detProc;
-        detectorModel = detMod;
-        console.log(
-          "[worker] Detector id2label:",
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          JSON.stringify((detMod.config as any)?.id2label ?? {}),
-        );
-        // Free any previously-loaded SAM 3 sessions if the user switched
-        // away from a text-promptable detector.
-        await unloadSam3();
-      }
+        },
+      });
 
       // Load classifier processor + model (WebGPU if available)
       const [clsProc, clsMod] = await Promise.all([
@@ -305,17 +220,13 @@ addEventListener("message", async (event: MessageEvent) => {
 
   // ── Run inference ────────────────────────────────────────────────────────
   if (data.type === "run-inference") {
-    // Validation differs by detector kind. SAM 3 doesn't use the
-    // transformers.js detectorModel/Processor pair — its readiness is
-    // managed inside the sam3 module — so the check is config-driven.
+    // Validation differs by detector kind — the detector module knows how to
+    // check readiness for each path (SAM 3 readiness is managed internally).
     if (!classifierModel || !classifierProcessor || !loadedConfig) {
       send({ type: "error", message: "Models not loaded" });
       return;
     }
-    if (
-      loadedConfig.detectorKind !== "text-promptable-segmentation" &&
-      (!detectorModel || !detectorProcessor)
-    ) {
+    if (!isDetectorReady(loadedConfig)) {
       send({ type: "error", message: "Detector not loaded" });
       return;
     }
@@ -337,145 +248,15 @@ addEventListener("message", async (event: MessageEvent) => {
         rawImage.height,
       );
 
-      // Detector inference — both branches produce `detections` + `labelForClass`
-      // so the box-building loop below can treat them uniformly.
-      let detections: PostProcessedDetection;
-      let labelForClass: (classIdx: number) => string;
-
-      if (config.detectorKind === "text-promptable-segmentation") {
-        // SAM 3 path — sam3 module handles preprocessing, inference, post-processing.
-        const prompt = data.prompt?.trim() || "seed";
-        console.log(
-          `[worker] Running SAM 3 detector with prompt: "${prompt}", threshold: ${config.detectorThreshold}`,
-        );
-        const sam3Result = await runSam3(
-          imageSrc,
-          prompt,
-          config.detectorThreshold,
-          rawImage.width,
-          rawImage.height,
-        );
-        detections = sam3Result;
-        // Open-vocabulary — every detection gets the prompt as its label.
-        labelForClass = () => prompt;
-        console.log(
-          `[worker] SAM 3 returned ${detections.boxes.length} detections`,
-        );
-      } else {
-        // Closed-vocabulary detector path (RT-DETR, DETR, etc.) — original
-        // transformers.js flow.
-        if (!detectorModel || !detectorProcessor) {
-          throw new Error("Detector model is null in non-SAM3 path");
-        }
-
-        const detInputs = await detectorProcessor(rawImage);
-        const detOutputs = await detectorModel(detInputs);
-        console.log("[worker] Detector output keys:", Object.keys(detOutputs));
-        for (const [key, val] of Object.entries(detOutputs)) {
-          const t = val as {
-            dims?: number[];
-            type?: string;
-            data?: Float32Array;
-          };
-          if (t?.dims) {
-            console.log(
-              `[worker]   ${key}: dims=${JSON.stringify(t.dims)} dtype=${t.type}`,
-            );
-          }
-          if (key === "logits" && t?.data) {
-            const scores = Array.from(t.data).map(
-              (v: number) => 1 / (1 + Math.exp(-v)),
-            ); // sigmoid
-            const sorted = [...scores].sort((a, b) => b - a);
-            console.log("[worker] Top 10 sigmoid scores:", sorted.slice(0, 10));
-            console.log(
-              "[worker] Scores > 0.01:",
-              scores.filter((s: number) => s > 0.01).length,
-            );
-          }
-        }
-
-        // Post-process detections
-        // RT-DETR uses sigmoid (no background class), so pass is_zero_shot=true
-        const numClasses = detOutputs.logits.dims[2];
-        const useSigmoid = numClasses === 1;
-
-        // Get boxes in model input space (640x640), then scale to original
-        // image dimensions ourselves — matching the Python CLI approach.
-        // post_process with null target_sizes returns normalized [0,1] boxes.
-        console.log(
-          "[worker] Post-processing with threshold:",
-          config.detectorThreshold,
-          "sigmoid:",
-          useSigmoid,
-        );
-        const processed = (
-          detectorProcessor.image_processor ?? detectorProcessor
-        ).post_process_object_detection(
-          detOutputs,
-          config.detectorThreshold,
-          null, // get normalized boxes
-          useSigmoid,
-        ) as PostProcessedDetection[];
-
-        // Scale normalized boxes from padded model space to original image coords.
-        // The model input is 640x640 (padded). The image was resized preserving
-        // aspect ratio, so we need to scale through the resized dimensions.
-        const modelW = detInputs.pixel_values.dims[3];
-        const modelH = detInputs.pixel_values.dims[2];
-        const resizeScale = Math.min(
-          modelW / rawImage.width,
-          modelH / rawImage.height,
-        );
-        const resizedW = rawImage.width * resizeScale;
-        const resizedH = rawImage.height * resizeScale;
-        const scaleX = rawImage.width / resizedW;
-        const scaleY = rawImage.height / resizedH;
-
-        console.log(
-          "[worker] Model input:",
-          modelW,
-          "x",
-          modelH,
-          "resized:",
-          resizedW.toFixed(0),
-          "x",
-          resizedH.toFixed(0),
-          "scale:",
-          scaleX.toFixed(3),
-          "x",
-          scaleY.toFixed(3),
-        );
-
-        // Convert normalized boxes to original image pixel coordinates
-        for (const det of processed) {
-          for (let i = 0; i < det.boxes.length; i++) {
-            const [x0, y0, x1, y1] = det.boxes[i];
-            det.boxes[i] = [
-              x0 * modelW * scaleX,
-              y0 * modelH * scaleY,
-              x1 * modelW * scaleX,
-              y1 * modelH * scaleY,
-            ];
-          }
-        }
-
-        console.log(
-          "[worker] Post-processed detections:",
-          JSON.stringify(processed),
-        );
-
-        const id2label = detectorModel.config?.id2label ?? {};
-        detections = processed[0];
-        labelForClass = (classIdx: number) =>
-          id2label[classIdx] ?? `class_${classIdx}`;
-        console.log(
-          "[worker] Detections count:",
-          detections?.boxes?.length ?? 0,
-          "id2label keys:",
-          Object.keys(id2label).length,
-        );
-      }
+      // Detector inference is delegated to the detector module. Both paths
+      // return `detections` + `labelForClass` so the box-building loop below
+      // can treat them uniformly.
+      const { detections, labelForClass } = await runDetector(
+        config,
+        imageSrc,
+        rawImage,
+        data.prompt,
+      );
 
       if (!detections || !detections.boxes || detections.boxes.length === 0) {
         console.log("[worker] No detections above threshold");
